Replace the timedelta test.
The timedelta test covered a specific issue with fields that didn't
round-trip their JSON identically.  That pulled in more code from
edx-platform than we'd like for a test.  This test shows the same
behavior, but in a much smaller test.

BTW: I don't think the field data machinery is still susceptible to this
problem, because it doesn't rely on json as it used to, but it's good to
keep the test cases we have.