Make Field.to_string() and Field.from_string() methods more consistent.
The old version of these methods was inherently assymmetric.  The deserialisation in from_string()
only called from_json() when enforce_type() was overridden to do so (which all classes with a
non-trivial from_json() implementation do) and enable_enforce_type was set to True.  This commit
makes the two functions more consistent by not considering enable_enforce_type at all for
deserialisation and calling from_json() in enforce_type() by default.  Serialisation and
deserialisation requires correct types to work already, so there is no harm in always enforciing
the type – the code fails in the same cases as before.

This commit fixes a bug in serialising DateTime fields.  The old version included spurious double
quotes around the string, which would not be correctly deserialised when enable_enforce_type was
not set (the default).