Add a way to stream unprotected local resources.
Depending on whether or not you're in a production environment, you may
want to be able to stream all resources from a given XBlock.  This
happens, in particular, when edx-platform is running in debug mode, and
all resources are required individually, instead of being bundled up.

Instead of requiring consumers to do their own pkg_resources lookups to
sidestep open_local_resource's 'public' restriction, we now have a
method that allows this to happen in a controlled way from the XBlock
itself, which takes into consideration things like a custom
resources_dir configuration, etc.