Make Mixologist able to handle class path strings. (#426)
Make Mixologist able to handle class path strings.

Previously you could only pass in class objects to the Mixologist.  This
meant that the mixin classes had to be imported somewhere else.  This
allows us to defer importing of mixin classes to when the Mixologist is
instantiated.

The reason for this need is that in edx-platform we currently import the
mixins as a part of the settings loading which casues issues in python
3. See BOM-66 for more details.