build: Don't explicityl pin the django version here.
It gets pinned in the common constraints file that gets pulled in and we
shouldn't pin it any further here.  This isn't used for unit testing but
is used for things like the docs builds so it should stay up-to-date but
is not critical to the functionality of the repo.