feat!: Assume that scope IDs are opaque key objects
BREAKING CHANGE: Raise an exception when scope IDs are missing or are not the
expected types. In particular, definition IDs must be DefinitionKey instances
and usage IDs must be UsageKey instances. This has been effectively true within
edx-platform (the lone production client of the XBlock library) for a long
time, but explictly enforcing it will now allow us to add strong type
annotations to XBlock in an upcoming release.

Bump version to 6.0.0.

Closes: https://github.com/openedx/XBlock/issues/708