feat: load_block_type_with_mixins
There are a few places in edx-platform where
we need the mixed block (or aside) class. These
places have to access the `mixologist` attribute on
the runtime, which seems more like an implementation
detail than an API. So, we add a couple simple
methods to achieve this in a more API-like way.